requirements.txt
Flask==2.0.2
python-dotenv==0.19.0
pandas==1.3.3
lxml==4.6.3
click==8.0.1
pytest==6.2.5
validators==0.18.2